Added More Features and fixed common bugs

+++ File organizer.py
@@ -1,6 +1,8 @@
 from tkinter import *
 import os
 import sys
+
+
 # from pathlib import path
 cwd = f"[{os.getcwd()}]"
 root = Tk()
@@ -30,6 +32,8 @@ label1 = Label(root, text="To organize file types into Folders Automatically",fo
 label2=Label(root,text="Current Folder "+cwd,bg="light cyan",fg="Black",font=("SegoeUI",9)).pack(side=TOP,pady=5)
 label3 = Label(root,text="Click To Organize Files",bg="Sky Blue",fg="blue",font=("Times new roman",13,"bold")).pack(side=TOP,pady=8)
 
+
+
 def allfiles():
     import os
     from pathlib import Path
@@ -37,7 +41,7 @@ def allfiles():
     "HTML": [".html5", ".html", ".htm", ".xhtml"],
 
     "IMAGES": [".jpeg", ".jpg", ".tiff", ".gif", ".bmp", ".png", ".bpg", ".svg",
-               ".heif", ".psd",".webp"],
+               ".heif", ".psd",".webp",".ico"],
 
     "VIDEOS": [".avi", ".flv", ".wmv", ".mov", ".mp4", ".webm", ".vob", ".mng",
                ".qt", ".mpg", ".3gp", ".m4v", ".mkv",".mpeg",".rm", ".swf"],
@@ -53,13 +57,18 @@ def allfiles():
     "AUDIO": [".aac", ".aa", ".aac", ".dvf", ".m4a", ".m4b", ".m4p", ".mp3",
               ".msv", ".ogg", ".oga", ".raw", ".vox", ".wav", ".wma"],
 
-    "PLAINTEXT": [".txt", ".in", ".out"],
+    "PLAINTEXT": [".txt", ".in", ".out",".md"],
 
     "PDF": [".pdf"],
     "PYTHON": [".py",".pyc"],
+    "C++":[".cpp"],
+    "Java":[".java"],
     "XML": [".xml"],
-    "EXE": [".exe"],
+    "EXE": [".exe",".msi"],
     "SHELL": [".sh"]}
+
+    EXCLUDE_FILE_NAME = "File Organizer.exe"
+
     FILE_FORMATS = {file_format: directory
                 for directory, file_formats in DIRECTORIES.items()
                 for file_format in file_formats}
@@ -69,6 +78,11 @@ def allfiles():
                 continue
             file_path = Path(entry)
             file_format = file_path.suffix.lower()
+            file_name = file_path.name
+
+            if file_name.lower() == EXCLUDE_FILE_NAME.lower():
+                continue
+
             if file_format in FILE_FORMATS:
                 directory_path = Path(FILE_FORMATS[file_format])
                 directory_path.mkdir(exist_ok=True)
@@ -87,7 +101,10 @@ def onlydoc():
     "DOCUMENTS": [".oxps", ".epub", ".pages", ".docx", ".doc", ".fdf", ".ods",
                   ".odt", ".pwi", ".xsn", ".xps", ".dotx", ".docm", ".dox",
                   ".rvg", ".rtf", ".rtfd", ".wpd", ".xls", ".xlsx", ".ppt",
-                  "pptx",".txt"] }
+                  "pptx",".txt",".pdf",".xml"] }
+
+    EXCLUDE_FILE_NAME = "File Organizer.exe"
+
     FILE_FORMATS = {file_format: directory
                 for directory, file_formats in DIRECTORIES.items()
                 for file_format in file_formats}
@@ -97,10 +114,16 @@ def onlydoc():
                 continue
             file_path = Path(entry)
             file_format = file_path.suffix.lower()
+            file_name = file_path.name
+
+            if file_name.lower() == EXCLUDE_FILE_NAME.lower():
+                continue
+
             if file_format in FILE_FORMATS:
                 directory_path = Path(FILE_FORMATS[file_format])
-                directory_path.mkdir(True, **('exist_ok',))
+                directory_path.mkdir(exist_ok=True)
                 file_path.rename(directory_path.joinpath(file_path))
+
             for dir in os.scandir():
                 try:
                     os.rmdir(dir)
@@ -113,7 +136,10 @@ def onlyimages():
     from pathlib import Path
     DIRECTORIES = {
     "IMAGES": [".jpeg", ".jpg", ".tiff", ".gif", ".bmp", ".png", ".bpg", ".svg",
-               ".heif",".heic", ".psd",".webp",".raw"]}
+               ".heif",".heic", ".psd",".webp",".raw",".ico"]}
+
+    EXCLUDE_FILE_NAME = "File Organizer.exe"
+
     FILE_FORMATS = {file_format: directory
                 for directory, file_formats in DIRECTORIES.items()
                 for file_format in file_formats}
@@ -123,9 +149,16 @@ def onlyimages():
                 continue
             file_path = Path(entry)
             file_format = file_path.suffix.lower()
+
+            file_name = file_path.name
+
+            if file_name.lower() == EXCLUDE_FILE_NAME.lower():
+                continue
+
+
             if file_format in FILE_FORMATS:
                 directory_path = Path(FILE_FORMATS[file_format])
-                directory_path.mkdir(True, **('exist_ok',))
+                directory_path.mkdir(exist_ok=True)
                 file_path.rename(directory_path.joinpath(file_path))
             for dir in os.scandir():
                 try:
@@ -140,6 +173,9 @@ def onlyvideos():
     DIRECTORIES = {
     "VIDEOS": [".avi", ".flv", ".wmv", ".mov", ".mp4", ".webm", ".vob", ".mng",
                ".qt", ".mpg", ".3gp", ".m4v", ".mkv",".mpeg",".rm", ".swf"]}
+
+    EXCLUDE_FILE_NAME = "File Organizer.exe"
+
     FILE_FORMATS = {file_format: directory
                 for directory, file_formats in DIRECTORIES.items()
                 for file_format in file_formats}
@@ -149,9 +185,16 @@ def onlyvideos():
                 continue
             file_path = Path(entry)
             file_format = file_path.suffix.lower()
+
+            file_name = file_path.name
+
+            if file_name.lower() == EXCLUDE_FILE_NAME.lower():
+                continue
+
+
             if file_format in FILE_FORMATS:
                 directory_path = Path(FILE_FORMATS[file_format])
-                directory_path.mkdir(True, **('exist_ok',))
+                directory_path.mkdir(exist_ok=True)
                 file_path.rename(directory_path.joinpath(file_path))
             for dir in os.scandir():
                 try:
@@ -167,6 +210,8 @@ def onlyaudios():
     "AUDIO": [".aac", ".aa", ".aac", ".dvf", ".m4a", ".m4b", ".m4p", ".mp3",
               ".msv", ".ogg", ".oga", ".raw", ".vox", ".wav", ".wma",".aiff", ".amr", ".flac", ".opus"]
               }
+    EXCLUDE_FILE_NAME = "File Organizer.exe"
+
     FILE_FORMATS = {file_format: directory
                 for directory, file_formats in DIRECTORIES.items()
                 for file_format in file_formats}
@@ -176,9 +221,15 @@ def onlyaudios():
                 continue
             file_path = Path(entry)
             file_format = file_path.suffix.lower()
+
+            file_name = file_path.name
+
+            if file_name.lower() == EXCLUDE_FILE_NAME.lower():
+                continue
+
             if file_format in FILE_FORMATS:
                 directory_path = Path(FILE_FORMATS[file_format])
-                directory_path.mkdir(True, **('exist_ok',))
+                directory_path.mkdir(exist_ok=True)
                 file_path.rename(directory_path.joinpath(file_path))
             for dir in os.scandir():
                 try:
